Move celestial bodies into new directory
